Changed line Encodings to CRLF and changed Budget to Decimal and fixed namespace issues

+++ Planner/User.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+
 namespace Planner
 {
     /// <summary>
@@ -6,8 +7,9 @@ namespace Planner
     /// </summary>
     public class User
     {
-        private float _budget;
-        public float Budget
+        private decimal _budget;
+
+        public decimal Budget
         {
             get;
             set;
@@ -17,6 +19,5 @@ namespace Planner
         {
             throw new NotImplementedException();
         }
-
     }
 }
